+++ src/collective/glossary/content.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
-from collective.glossary.interfaces import IGlossary, ITerm
-from plone.dexterity.content import Container, Item
+from collective.glossary.interfaces import IGlossary
+from collective.glossary.interfaces import ITerm
+from plone.dexterity.content import Container
+from plone.dexterity.content import Item
 from zope.interface import implementer